+++ backend/src/domain/entities/business/payment.ts
@@ -1,12 +1,14 @@
-import { User } from './user/user';
+import { User } from '../user/user';
 import { Work } from './work';
-import { Business } from './business/business';
+import { Business } from './business';
 
-export class Review {
+export class Payment {
   public id: string;
-  public rate: number;
-  public comment?: string;
-  public reviewedAt: Date;
+  public amount: number;
+  public status: string;
+  public paymentMethod: string;
+  public transactionId?: string;
+  public timestamp: Date;
   public workId: string;
   public work?: Work;
   public clientId: string;
@@ -15,7 +17,7 @@ export class Review {
   public business?: Business;
 
   /**
-   * Updates the review with the provided data.
+   * Updates the payment with the provided data.
    *
    * @param partial - A partial object containing the updated data. The following
    *                  properties are not allowed and will be deleted if present:
@@ -24,7 +26,7 @@ export class Review {
    *                    - clientId
    *                    - businessId
    */
-  update(partial: Partial<Review>) {
+  update(partial: Partial<Payment>) {
     delete partial.id;
     delete partial.workId;
     delete partial.clientId;
