+++ backend/src/domain/entities/index.ts
@@ -1,10 +1,8 @@
 export * from './account';
 export * from './appointment';
-export * from './availability';
 export * from './business/business';
 export * from './category';
-export * from './payment';
-export * from './review';
+export * from './business/payment';
+export * from './business/review';
 export * from './token-version';
 export * from './verification';
-export * from './work';
